requirements.txt
absl-py==2.0.0
annotated-types==0.6.0
ansi2html==1.8.0
anyio==4.2.0
appdirs==1.4.4
appnope==0.1.3
asttokens==2.4.0
astunparse==1.6.3
az==0.1.0.dev1
backcall==0.2.0
beautifulsoup4==4.12.2
cachetools==5.3.2
certifi==2023.7.22
charset-normalizer==3.3.1
click==8.1.7
comm==0.1.4
contourpy==1.1.0
cycler==0.11.0
dash==2.14.0
dash-core-components==2.0.0
dash-daq==0.5.0
dash-html-components==2.0.0
dash-table==5.0.0
debugpy==1.8.0
decorator==5.1.1
distlib==0.3.7
distro==1.8.0
executing==2.0.0
filelock==3.12.4
Flask==2.2.5
flatbuffers==23.5.26
fonttools==4.42.0
frozendict==2.3.8
gast==0.5.4
gh==0.0.4
gitdb==4.0.11
GitPython==3.1.40
google-auth==2.26.1
google-auth-oauthlib==1.2.0
google-pasta==0.2.0
grpcio==1.60.0
gunicorn==21.2.0
h11==0.14.0
h5py==3.10.0
html5lib==1.1
httpcore==1.0.2
httpx==0.26.0
idna==3.4
importlib-metadata==6.8.0
iniconfig==2.0.0
ipykernel==6.25.2
ipython==8.16.1
itsdangerous==2.1.2
jedi==0.19.1
Jinja2==3.1.2
joblib==1.3.2
jupyter_client==8.3.1
jupyter_core==5.4.0
keras==2.15.0
kiwisolver==1.4.4
libclang==16.0.6
lxml==4.9.3
Markdown==3.5.1
MarkupSafe==2.1.3
matplotlib==3.8.2
matplotlib-inline==0.1.6
ml-dtypes==0.2.0
multitasking==0.0.11
nest-asyncio==1.5.8
nltk==3.8.1
numpy==1.25.2
oauthlib==3.2.2
openai==1.6.1
opt-einsum==3.3.0
packaging==23.2
pandas==2.0.3
pandas-datareader==0.10.0
parso==0.8.3
patsy==0.5.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==10.0.0
platformdirs==3.11.0
plotly==5.17.0
plotly-express==0.4.1
pluggy==1.3.0
progressbar2==4.3.2
prompt-toolkit==3.0.39
protobuf==4.23.4
psutil==5.9.5
ptyprocess==0.7.0
pure-eval==0.2.2
pyasn1==0.5.1
pyasn1-modules==0.3.0
pydantic==2.5.3
pydantic_core==2.14.6
pygam==0.9.0
Pygments==2.16.1
PyMySQL==1.1.0
pyparsing==3.0.9
pytest==7.4.3
python-dateutil==2.8.2
python-dotenv==1.0.1
python-utils==3.8.1
pytz==2023.3
pyzmq==25.1.1
regex==2023.12.25
requests==2.31.0
requests-oauthlib==1.3.1
retrying==1.3.4
rsa==4.9
scikit-learn==1.3.2
scipy==1.11.4
six==1.16.0
smmap==5.0.1
sniffio==1.3.0
soupsieve==2.4.1
SQLAlchemy==2.0.27
stack-data==0.6.3
statsmodels==0.14.0
tenacity==8.2.3
tensorboard==2.15.1
tensorboard-data-server==0.7.2
tensorflow==2.15.0
tensorflow-estimator==2.15.0
tensorflow-io-gcs-filesystem==0.34.0
tensorflow-macos==2.15.0
termcolor==2.4.0
threadpoolctl==3.2.0
tornado==6.3.3
tqdm==4.66.1
traitlets==5.11.2
typing_extensions==4.8.0
tzdata==2023.3
urllib3==2.0.7
virtualenv==20.24.6
wcwidth==0.2.8
webencodings==0.5.1
Werkzeug==2.2.3
wrapt==1.14.1
yfinance==0.2.26
zipp==3.17.0
